+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [next]
 
 - fix(): avoid resetting Group transform by default [#9525](https://github.com/fabricjs/fabric.js/pull/9525)
+- test() Backport a test to capture a failing text style situation [#9531](https://github.com/fabricjs/fabric.js/pull/9531)
 
 ## [6.0.0-beta16]